Backend/requirements.txt
earthengine-api>=0.1.350
geemap>=0.35.3
geopandas>=0.13.0
rasterio>=1.3.0
matplotlib>=3.7.0
folium>=0.14.0
requests>=2.31.0
numpy>=1.24.0
pandas>=2.0.0
pyyaml>=6.0
jupyter>=1.0.0
notebook>=6.5.0
scikit-learn>=1.0.0
scipy>=1.7.0
matplotlib>=3.5.0
seaborn>=0.11.0
plotly>=5.0.0